Repair tests
