+++ triton-vm/src/vm.rs
@@ -247,7 +247,7 @@ mod triton_vm_tests {
     use super::*;
     use crate::instruction::sample_programs;
     use crate::table::base_matrix::ProcessorMatrixRow;
-    use crate::table::base_table::{BaseTable, BaseTableTrait, HasBaseTable};
+    use crate::table::base_table::{BaseTableTrait, HasBaseTable};
     use crate::table::challenges_endpoints::{AllChallenges, AllEndpoints};
     use crate::table::extension_table::ExtensionTable;
     use crate::table::processor_table::ProcessorTable;
